requirements.txt
paho-mqtt>=1.6.1
requests>=2.28.0
sseclient-py>=1.7.2
